Reorganize css file structure
