+++ src/assets/css/index.css
@@ -1,91 +1,18 @@
-.Header {
-  display: flex;
-  justify-content: space-between;
-  padding: 40px;
-}
-
-.App {
-  background: yellowgreen;
-  padding: 8%;
-  color: white;
-  text-align: center;
-}
-
-.App-content {
-  width: 45%;
-  margin: auto;
-}
-
-.App-content p {
-  font-size: 14px;
-  line-height: 20px;
-}
-
-.App h3 {
-  border-bottom: 0.2px solid wheat;
-  padding-bottom: 20px;
-  font-weight: bolder;
-}
-
-.form select,
-.form .button {
-  border-radius: 20px;
-  padding: 10px;
-  margin: 10px;
-  width: 20%;
-}
-
-.form .button {
-  background: white;
-  color: yellowgreen;
-  border: none;
-}
-
-.form select {
-  background: none;
-  color: white;
-  border: 2px solid white;
-}
-
-.form select option {
-  background: white;
-  color: yellowgreen;
-}
-
-.RoomList {
-  margin: auto;
-  width: 90%;
-}
-
-.RoomList-text {
-  text-align: center;
-}
-
-.RoomList-content {
-  display: flex;
-  margin-top: 100px;
-}
-
-.RoomLDetails {
-  display: flex;
-  justify-content: space-between;
-  width: 60%;
-  margin: 0 0 0 300px;
-}
-
-.details-last p {
-  font-size: 9px;
-  font-weight: bolder;
-}
-
-.details-p p:nth-child(odd) {
-  background-color: #f2f2f2;
-  padding: 7px;
-}
-
-.details-last {
-  text-align: right;
-  margin-top: 30px;
+body {
+  height: 100vh;
+  margin: 0;
+  font-family:
+    'Segoe UI',
+    'Roboto',
+    'Oxygen',
+    'Ubuntu',
+    'Cantarell',
+    'Fira Sans',
+    'Droid Sans',
+    'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
 }
 
 /* Set a style for all buttons */
@@ -108,71 +35,8 @@ button {
   margin-top: 40px;
 }
 
-.back {
-  margin-left: 180px;
-}
-
-.bck {
-  width: 10%;
-}
-
-.SideBar h3 {
-  text-align: center;
-}
-
-.SideBar ul {
-  margin-top: 100px;
-}
-
-.SideBar ul li {
-  list-style: none;
-  padding: 10px;
-}
-
-.active a {
-  color: white;
-}
-
-.SideBar ul li a {
-  text-decoration: none;
-  font-weight: bolder;
-  color: black;
-}
-
-.active {
-  background-color: yellowgreen;
-}
-
-.SideBar-social {
-  margin-top: 340px;
-  text-align: center;
-}
-
-.Room-img img {
-  width: 95%;
-}
-
-.Room-details-img img {
-  width: 70%;
-}
-
-.description {
-  color: lightgray;
-  font-size: 14px;
-  margin-top: 10px;
-}
-
-.title {
-  border-bottom: 3px dotted lightgray;
-  padding-bottom: 17px;
-  width: 13%;
-}
-
-
-
 /* login and signup forms */
 
-/* Full-width input fields */
 input[type=text],
 input[type=password],
 input[type=date],
@@ -282,19 +146,4 @@ span.psw {
   .cancelbtn {
     width: 100%;
   }
-}
-
-
-/* reservation styles*/
-.ReservationList {
-  margin: auto;
-  width: 95%;
-}
-.ReservationList-content {
-  display: grid;
-  grid-template-columns: auto auto auto auto;
-}
-
-.Reservation-img img {
-  width: 50%;
 }
