+++ src/components/App.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
+import '../assets/css/app.css'
 import LoginForm from './forms/LoginForm'
 import SignupForm from './forms/SignupForm'
-import Modal from "./Modal"
+import Modal from "./modals/Modal"
 import LoginModal from './modals/LoginModal'
 import { close, show, close_login, show_login } from '../actions/index'
