+++ src/index.js
@@ -5,6 +5,7 @@ import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 import './assets/css/index.css';
+import './assets/css/sidebar.css';
 import rootReducer from '../src/reducers/index'
 import Router from './components/Router';
